+++ headers/classes/school.h
@@ -9,12 +9,12 @@ namespace lsim {
 	class School : public lsim::Occupation {
 		public:
 			School(lsim::mainCharacter *nSelf, int nIndex, int nSchoolType = lsim::SCHOOLNULL, int nCurrentYear = 0);
-			int updateGrades();
+			short int updateGrades();
 			void goToMenu();
 			void passAYear();
 		private:
-			int currentYear;
-			int avgGrades;
+			short int currentYear;
+			short int avgGrades;
 			lsim::Teacher teacher;
 			std::vector<lsim::Classmate> classmates;
 			lsim::io::Menu classmatesMenu;
